Move drunet utilities to recon module.
